+++ src/utils.ts
@@ -30,6 +30,7 @@ import {
 import { Mjolnir } from "./Mjolnir";
 import { ClientRequest, IncomingMessage } from "http";
 import { default as parseDuration } from "parse-duration";
+import ManagementRoomOutput from "./ManagementRoom";
 
 // Define a few aliases to simplify parsing durations.
 
@@ -70,17 +71,17 @@ export function isTrueJoinEvent(event: any): boolean {
     return membership === 'join' && prevMembership !== "join";
 }
 
-export async function redactUserMessagesIn(mjolnir: Mjolnir, userIdOrGlob: string, targetRoomIds: string[], limit = 1000) {
+export async function redactUserMessagesIn(client: MatrixClient, managementRoom: ManagementRoomOutput, userIdOrGlob: string, targetRoomIds: string[], limit = 1000, noop = false) {
     for (const targetRoomId of targetRoomIds) {
-        await mjolnir.logMessage(LogLevel.DEBUG, "utils#redactUserMessagesIn", `Fetching sent messages for ${userIdOrGlob} in ${targetRoomId} to redact...`, targetRoomId);
+        await managementRoom.logMessage(LogLevel.DEBUG, "utils#redactUserMessagesIn", `Fetching sent messages for ${userIdOrGlob} in ${targetRoomId} to redact...`, targetRoomId);
 
-        await getMessagesByUserIn(mjolnir.client, userIdOrGlob, targetRoomId, limit, async (eventsToRedact) => {
+        await getMessagesByUserIn(client, userIdOrGlob, targetRoomId, limit, async (eventsToRedact) => {
             for (const victimEvent of eventsToRedact) {
-                await mjolnir.logMessage(LogLevel.DEBUG, "utils#redactUserMessagesIn", `Redacting ${victimEvent['event_id']} in ${targetRoomId}`, targetRoomId);
-                if (!mjolnir.config.noop) {
-                    await mjolnir.client.redactEvent(targetRoomId, victimEvent['event_id']);
+                await managementRoom.logMessage(LogLevel.DEBUG, "utils#redactUserMessagesIn", `Redacting ${victimEvent['event_id']} in ${targetRoomId}`, targetRoomId);
+                if (!noop) {
+                    await client.redactEvent(targetRoomId, victimEvent['event_id']);
                 } else {
-                    await mjolnir.logMessage(LogLevel.WARN, "utils#redactUserMessagesIn", `Tried to redact ${victimEvent['event_id']} in ${targetRoomId} but Mjolnir is running in no-op mode`, targetRoomId);
+                    await managementRoom.logMessage(LogLevel.WARN, "utils#redactUserMessagesIn", `Tried to redact ${victimEvent['event_id']} in ${targetRoomId} but Mjolnir is running in no-op mode`, targetRoomId);
                 }
             }
         });
@@ -187,51 +188,6 @@ export async function getMessagesByUserIn(client: MatrixClient, sender: string, 
             return;
         }
     } while (token && processed < limit)
-}
-
-/*
- * Take an arbitrary string and a set of room IDs, and return a
- * TextualMessageEventContent whose plaintext component replaces those room
- * IDs with their canonical aliases, and whose html component replaces those
- * room IDs with their matrix.to room pills.
- *
- * @param client The matrix client on which to query for room aliases
- * @param text An arbitrary string to rewrite with room aliases and pills
- * @param roomIds A set of room IDs to find and replace in `text`
- * @param msgtype The desired message type of the returned TextualMessageEventContent
- * @returns A TextualMessageEventContent with replaced room IDs
- */
-export async function replaceRoomIdsWithPills(mjolnir: Mjolnir, text: string, roomIds: Set<string>, msgtype: MessageType = "m.text"): Promise<TextualMessageEventContent> {
-    const content: TextualMessageEventContent = {
-        body: text,
-        formatted_body: htmlEscape(text),
-        msgtype: msgtype,
-        format: "org.matrix.custom.html",
-    };
-
-    const escapeRegex = (v: string): string => {
-        return v.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
-    };
-
-    const viaServers = [(new UserID(await mjolnir.client.getUserId())).domain];
-    for (const roomId of roomIds) {
-        let alias = roomId;
-        try {
-            alias = (await mjolnir.client.getPublishedAlias(roomId)) || roomId;
-        } catch (e) {
-            // This is a recursive call, so tell the function not to try and call us
-            await mjolnir.logMessage(LogLevel.WARN, "utils", `Failed to resolve room alias for ${roomId} - see console for details`, null, true);
-            LogService.warn("utils", extractRequestError(e));
-        }
-        const regexRoomId = new RegExp(escapeRegex(roomId), "g");
-        content.body = content.body.replace(regexRoomId, alias);
-        if (content.formatted_body) {
-            const permalink = Permalinks.forRoom(alias, alias !== roomId ? [] : viaServers);
-            content.formatted_body = content.formatted_body.replace(regexRoomId, `<a href="${permalink}">${alias}</a>`);
-        }
-    }
-
-    return content;
 }
 
 let isMatrixClientPatchedForConciseExceptions = false;
