+++ test/integration/utilsTest.ts
@@ -1,15 +1,11 @@
 import { strict as assert } from "assert";
-
-import { UserID } from "matrix-bot-sdk";
-import config from "../../src/config";
-import { replaceRoomIdsWithPills } from "../../src/utils";
+import { LogLevel } from "matrix-bot-sdk";
+import ManagementRoomOutput from "../../src/ManagementRoom";
 
 describe("Test: utils", function() {
     it("replaceRoomIdsWithPills correctly turns a room ID in to a pill", async function() {
-        this.timeout(20000);
-
         const managementRoomAlias = this.config.managementRoom;
-
+        const managementRoomOutput: ManagementRoomOutput = this.mjolnir.managementRoom;
         await this.mjolnir.client.sendStateEvent(
             this.mjolnir.managementRoomId,
             "m.room.canonical_alias",
@@ -17,15 +13,20 @@ describe("Test: utils", function() {
             { alias: managementRoomAlias }
         );
 
-        const out = await replaceRoomIdsWithPills(
-            this.mjolnir,
-            `it's fun here in ${this.mjolnir.managementRoomId}`,
-            new Set([this.mjolnir.managementRoomId, "!myfaketestid:example.com"])
-        );
-
-        const ourHomeserver = new UserID(await this.mjolnir.client.getUserId()).domain;
+        const message: any = await new Promise(async resolve => {
+            this.mjolnir.client.on('room.message', (roomId, event) => {
+                if (roomId === this.mjolnir.managementRoomId) {
+                    if (event.content?.body?.startsWith("it's")) {
+                        resolve(event);
+                    }
+                }
+            })
+            await managementRoomOutput.logMessage(LogLevel.INFO, 'replaceRoomIdsWithPills test',
+                `it's fun here in ${this.mjolnir.managementRoomId}`,
+                [this.mjolnir.managementRoomId, "!myfaketestid:example.com"]);
+        });
         assert.equal(
-            out.formatted_body,
+            message.content.formatted_body,
             `it's fun here in <a href="https://matrix.to/#/${managementRoomAlias}">${managementRoomAlias}</a>`
         );
     });
